Issue/847  paged attention prefill一段式接口

+++ src/infiniop/ops/paged_attention_prefill/cuda/kernel.cuh
@@ -4,10 +4,10 @@
 namespace op::paged_attention_prefill::cuda {
 
 // 辅助函数：二分查找确定当前 global_token_idx 属于哪个 sequence
-__device__ __forceinline__ int find_seq_id(int token_idx, const int64_t *offset, int num_seqs) {
-    int low = 0, high = num_seqs - 1;
+__device__ __forceinline__ size_t find_seq_id(size_t token_idx, const int64_t *offset, size_t num_seqs) {
+    size_t low = 0, high = num_seqs - 1;
     while (low <= high) {
-        int mid = (low + high) >> 1;
+        size_t mid = (low + high) >> 1;
         if (token_idx >= offset[mid] && token_idx < offset[mid + 1]) {
             return mid;
         } else if (token_idx < offset[mid]) {
@@ -32,22 +32,22 @@ __global__ void pagedAttentionPrefillKernel(
     const size_t num_seqs) {
 
     // --- 使用 2D Grid 坐标 ---
-    const int global_token_idx = blockIdx.x; // 展平后的全局 token 索引
-    const int head_idx = blockIdx.y;         // Head 索引
-    const int dim_idx = threadIdx.x;         // Head 内部维度
+    const size_t global_token_idx = blockIdx.x; // 展平后的全局 token 索引
+    const size_t head_idx = blockIdx.y;         // Head 索引
+    const size_t dim_idx = threadIdx.x;         // Head 内部维度
 
     if (dim_idx >= head_size) {
         return;
     }
 
     // --- 通过二分查找 offset 找到所属的 seq_idx ---
-    int seq_idx = find_seq_id(global_token_idx, offset_, num_seqs);
+    size_t seq_idx = find_seq_id(global_token_idx, offset_, num_seqs);
 
     // --- 获取该 Sequence 本次 Prefill 的长度
     const int64_t cur_new_len = seq_lens_[seq_idx];
 
     // --- 该 token 在当前序列中的相对位置
-    int q_token_idx = global_token_idx - offset_[seq_idx];
+    size_t q_token_idx = global_token_idx - offset_[seq_idx];
 
     const Tdata *q_ptr_base = q_ + global_token_idx * num_heads * head_size + head_idx * head_size;
     Tdata *out_ptr = out_ + global_token_idx * num_heads * head_size + head_idx * head_size;
@@ -65,14 +65,14 @@ __global__ void pagedAttentionPrefillKernel(
 
     // Pass 1: 计算 Score 并找最大值
     Tcompute max_score = -FLT_MAX;
-    for (int t = 0; t <= causal_limit; ++t) {
+    for (size_t t = 0; t <= causal_limit; ++t) {
         const int64_t b_idx = t / block_size;
         const int64_t t_off = t % block_size;
         const int64_t physical_block_id = block_table[b_idx];
         const Tdata *k_vec = k_cache_ + physical_block_id * kv_block_stride + kv_head_idx * kv_head_stride + t_off * head_size;
 
         Tcompute score = 0.0f;
-        for (int d = 0; d < head_size; ++d) {
+        for (size_t d = 0; d < head_size; ++d) {
             score += static_cast<Tcompute>(q_ptr_base[d]) * static_cast<Tcompute>(k_vec[d]);
         }
         score *= static_cast<Tcompute>(scale);
@@ -86,14 +86,14 @@ __global__ void pagedAttentionPrefillKernel(
 
     // Pass 2: 计算 Sum of Exp
     Tcompute sum_exp = 0.0f;
-    for (int t = 0; t <= causal_limit; ++t) {
+    for (size_t t = 0; t <= causal_limit; ++t) {
         const int64_t b_idx = t / block_size;
         const int64_t t_off = t % block_size;
         const int64_t physical_block_id = block_table[b_idx];
         const Tdata *k_vec = k_cache_ + physical_block_id * kv_block_stride + kv_head_idx * kv_head_stride + t_off * head_size;
 
         Tcompute score = 0.0f;
-        for (int d = 0; d < head_size; ++d) {
+        for (size_t d = 0; d < head_size; ++d) {
             score += static_cast<Tcompute>(q_ptr_base[d]) * static_cast<Tcompute>(k_vec[d]);
         }
         score *= static_cast<Tcompute>(scale);
@@ -106,14 +106,14 @@ __global__ void pagedAttentionPrefillKernel(
     // Pass 3: 加权求和得到输出
     Tcompute acc = 0.0f;
     Tcompute inv_sum = 1.0f / (sum_exp + 1e-6f);
-    for (int t = 0; t <= causal_limit; ++t) {
+    for (size_t t = 0; t <= causal_limit; ++t) {
         const int64_t b_idx = t / block_size;
         const int64_t t_off = t % block_size;
         const int64_t physical_block_id = block_table[b_idx];
 
         const Tdata *k_vec = k_cache_ + physical_block_id * kv_block_stride + kv_head_idx * kv_head_stride + t_off * head_size;
         Tcompute score = 0.0f;
-        for (int d = 0; d < head_size; ++d) {
+        for (size_t d = 0; d < head_size; ++d) {
             score += static_cast<Tcompute>(q_ptr_base[d]) * static_cast<Tcompute>(k_vec[d]);
         }
         score *= static_cast<Tcompute>(scale);
